[FIX] tests: let browser_js's timeout govern the ready phase too
_wait_ready hardcoded timeout=60 and browser_js called it without an argument,
so the caller's budget only ever reached _wait_code_ok. A test that raised the
timeout for a slow ready expression still got 60s, and debug=True -- which sets
timeout to 1e6 for the express purpose of letting a human step through -- gave
up on the ready phase after 60s anyway.

Pass it through. browser_js already defaults to 60, the same value _wait_ready
defaulted to, so a default call is unchanged; only callers that override the
timeout, and debug/watch mode, see a difference. The annotation widens to float
because debug mode passes 1e6.

The two phases keep separate budgets, which is what architecture.md describes:
its "remaining budget" note is about the evaluate call inside _wait_code_ok, not
about sharing one budget with the ready phase.

Task ID: 25800

+++ odoo/tests/browser.py
@@ -800,7 +800,7 @@ which leads to stray network requests and inconsistencies."""
         params["name"] = name
         self._websocket_request("Network.deleteCookies", params=params)
 
-    def _wait_ready(self, ready_code: str | None = None, timeout: int = 60) -> bool:
+    def _wait_ready(self, ready_code: str | None = None, timeout: float = 60) -> bool:
         timeout *= self.throttling_factor
         ready_code = ready_code or "document.readyState === 'complete'"
         self._logger.info('Evaluate ready code "%s"', ready_code)
